refactor: renamed functions and moved kurtosis commands to new file
